移除多余文件
移除多余文件

+++ IceInk/Tools/00.CsharpExtensions/IOExtension.cs
@@ -22,13 +22,13 @@ using System.Linq;
 namespace IceInk
 {
     /// <summary>
-    /// 对 System.IO 的一些扩展
+    ///     对 System.IO 的一些扩展
     /// </summary>
     public static class IOExtension
     {
         /// <summary>
-        /// 创建新的文件夹,如果存在则不创建
-        ///  <code>
+        ///     创建新的文件夹,如果存在则不创建
+        ///     <code>
         ///  示例
         /// var testDir = "Assets/TestFolder";
         /// testDir.CreateDirIfNotExists();
@@ -37,17 +37,14 @@ namespace IceInk
         /// </summary>
         public static string CreateDirIfNotExists(this string dirFullPath)
         {
-            if (!Directory.Exists(dirFullPath))
-            {
-                Directory.CreateDirectory(dirFullPath);
-            }
+            if (!Directory.Exists(dirFullPath)) Directory.CreateDirectory(dirFullPath);
 
             return dirFullPath;
         }
 
         /// <summary>
-        /// 删除文件夹，如果存在
-        /// <code> 示例
+        ///     删除文件夹，如果存在
+        ///     <code> 示例
         /// var testDir = "Assets/TestFolder";
         /// testDir.DeleteDirIfExists();
         /// // 结果为，在 Assets 目录下删除了 TestFolder
@@ -55,15 +52,12 @@ namespace IceInk
         /// </summary>
         public static void DeleteDirIfExists(this string dirFullPath)
         {
-            if (Directory.Exists(dirFullPath))
-            {
-                Directory.Delete(dirFullPath, true);
-            }
+            if (Directory.Exists(dirFullPath)) Directory.Delete(dirFullPath, true);
         }
 
         /// <summary>
-        /// 清空 Dir（保留目录),如果存在。
-        /// <code>
+        ///     清空 Dir（保留目录),如果存在。
+        ///     <code>
         /// var testDir = "Assets/TestFolder";
         /// testDir.EmptyDirIfExists();
         /// // 结果为，清空了 TestFolder 里的内容
@@ -71,17 +65,14 @@ namespace IceInk
         /// </summary>
         public static void EmptyDirIfExists(this string dirFullPath)
         {
-            if (Directory.Exists(dirFullPath))
-            {
-                Directory.Delete(dirFullPath, true);
-            }
+            if (Directory.Exists(dirFullPath)) Directory.Delete(dirFullPath, true);
 
             Directory.CreateDirectory(dirFullPath);
         }
 
         /// <summary>
-        /// 删除文件 如果存在
-        /// <code>
+        ///     删除文件 如果存在
+        ///     <code>
         /// // 示例
         /// var filePath = "Assets/Test.txt";
         /// File.Create("Assets/Test);
@@ -96,12 +87,11 @@ namespace IceInk
             if (!File.Exists(fileFullPath)) return false;
             File.Delete(fileFullPath);
             return true;
-
         }
 
         /// <summary>
-        /// 合并路径
-        /// <code>
+        ///     合并路径
+        ///     <code>
         /// // 示例：
         /// Application.dataPath.CombinePath("Resources").LogInfo();  // /projectPath/Assets/Resources
         /// </code>
@@ -113,10 +103,10 @@ namespace IceInk
         {
             return Path.Combine(selfPath, toCombinePath);
         }
-        
+
 
         /// <summary>
-        /// 读取文本
+        ///     读取文本
         /// </summary>
         /// <param name="fileFullPath"></param>
         /// <returns></returns>
@@ -151,30 +141,30 @@ namespace IceInk
 #endif
 
         /// <summary>
-        /// 获取文件夹名
+        ///     获取文件夹名
         /// </summary>
         /// <param name="fileName"></param>
         /// <returns></returns>
         public static string GetDirectoryName(string fileName)
         {
-            fileName = IOExtension.MakePathStandard(fileName);
+            fileName = MakePathStandard(fileName);
             return fileName.Substring(0, fileName.LastIndexOf('/'));
         }
 
         /// <summary>
-        /// 获取文件名
+        ///     获取文件名
         /// </summary>
         /// <param name="path"></param>
         /// <param name="separator"></param>
         /// <returns></returns>
         public static string GetFileName(string path, char separator = '/')
         {
-            path = IOExtension.MakePathStandard(path);
+            path = MakePathStandard(path);
             return path.Substring(path.LastIndexOf(separator) + 1);
         }
 
         /// <summary>
-        /// 获取不带后缀的文件名
+        ///     获取不带后缀的文件名
         /// </summary>
         /// <param name="fileName"></param>
         /// <param name="separator"></param>
@@ -185,7 +175,7 @@ namespace IceInk
         }
 
         /// <summary>
-        /// 获取不带后缀的文件路径
+        ///     获取不带后缀的文件路径
         /// </summary>
         /// <param name="fileName"></param>
         /// <returns></returns>
@@ -195,49 +185,40 @@ namespace IceInk
         }
 
         /// <summary>
-        /// 使目录存在,Path可以是目录名必须是文件名
+        ///     使目录存在,Path可以是目录名必须是文件名
         /// </summary>
         /// <param name="path"></param>
         public static void MakeFileDirectoryExist(string path)
         {
-            string root = Path.GetDirectoryName(path);
-            if (!Directory.Exists(root))
-            {
-                Directory.CreateDirectory(root);
-            }
+            var root = Path.GetDirectoryName(path);
+            if (!Directory.Exists(root)) Directory.CreateDirectory(root);
         }
 
         /// <summary>
-        /// 使目录存在
+        ///     使目录存在
         /// </summary>
         /// <param name="path"></param>
         public static void MakeDirectoryExist(string path)
         {
-            if (!Directory.Exists(path))
-            {
-                Directory.CreateDirectory(path);
-            }
+            if (!Directory.Exists(path)) Directory.CreateDirectory(path);
         }
 
         /// <summary>
-        /// 结合目录
+        ///     结合目录
         /// </summary>
         /// <param name="paths"></param>
         /// <returns></returns>
         public static string Combine(params string[] paths)
         {
-            string result = string.Empty;
-            foreach (string path in paths)
-            {
-                result = Path.Combine(result, path);
-            }
+            var result = string.Empty;
+            foreach (var path in paths) result = Path.Combine(result, path);
 
             result = MakePathStandard(result);
             return result;
         }
 
         /// <summary>
-        /// 获取父文件夹
+        ///     获取父文件夹
         /// </summary>
         /// <param name="path"></param>
         /// <returns></returns>
@@ -248,7 +229,7 @@ namespace IceInk
 
 
         /// <summary>
-        /// 使路径标准化，去除空格并将所有'\'转换为'/'
+        ///     使路径标准化，去除空格并将所有'\'转换为'/'
         /// </summary>
         /// <param name="path"></param>
         /// <returns></returns>
@@ -263,21 +244,14 @@ namespace IceInk
             var pathList = new List<string>();
             var di = new DirectoryInfo(dirABSPath);
 
-            if (!di.Exists)
-            {
-                return pathList;
-            }
+            if (!di.Exists) return pathList;
 
             var files = di.GetFiles();
             foreach (var fi in files)
             {
                 if (!string.IsNullOrEmpty(suffix))
-                {
-                    if (!fi.FullName.EndsWith(suffix, System.StringComparison.CurrentCultureIgnoreCase))
-                    {
+                    if (!fi.FullName.EndsWith(suffix, StringComparison.CurrentCultureIgnoreCase))
                         continue;
-                    }
-                }
 
                 pathList.Add(fi.FullName);
             }
@@ -285,10 +259,7 @@ namespace IceInk
             if (isRecursive)
             {
                 var dirs = di.GetDirectories();
-                foreach (var d in dirs)
-                {
-                    pathList.AddRange(GetDirSubFilePathList(d.FullName, isRecursive, suffix));
-                }
+                foreach (var d in dirs) pathList.AddRange(GetDirSubFilePathList(d.FullName, isRecursive, suffix));
             }
 
             return pathList;
@@ -323,10 +294,7 @@ namespace IceInk
         {
             var lastIndex = absOrAssetsPath.LastIndexOf(".");
 
-            if (lastIndex >= 0)
-            {
-                return absOrAssetsPath.Substring(lastIndex);
-            }
+            if (lastIndex >= 0) return absOrAssetsPath.Substring(lastIndex);
 
             return string.Empty;
         }
@@ -345,7 +313,5 @@ namespace IceInk
 
             return absOrAssetsPath.EndsWith("/") ? dirs[dirs.Length - 2] : dirs[dirs.Length - 1];
         }
-
     }
-
 }
